requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
docx-mailmerge==0.5.0
lxml==4.6.2
pytz==2021.1
sqlparse==0.4.1
